requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl

boto3==1.26.6
botocore==1.29.6
Flask==2.2.2
matplotlib==3.6.2
pandas==1.5.1
Pillow==9.3.0
psutil==5.9.4
requests==2.28.1
torch==1.13.0
opencv-python
ipython
pyyaml
torchvision==0.14.0
tqdm==4.64.1
seaborn==0.12.1
scipy==1.9.3


